Conflicts:
	Data/level_1/cameras.xml
	MasterUAB/Core/Engine/UABEngine.cpp
	MasterUAB/Core/Engine/UABEngine.h
	MasterUAB/Graphics/Camera/Camera.h
	MasterUAB/Graphics/Camera/CameraController.h
	MasterUAB/Graphics/Camera/CameraControllerManager.cpp
	MasterUAB/Graphics/Camera/CameraControllerManager.h
	MasterUAB/Graphics/Camera/CameraInfo.cpp
	MasterUAB/Graphics/Camera/CameraKeyController.cpp
	MasterUAB/Graphics/Graphics.vcxproj.filters
	MasterUAB/VideogameTest/Application.cpp

JOAN NO LA LIES!

+++ MasterUAB/Graphics/RenderableObjects/RenderableObjectManager.cpp
@@ -7,14 +7,14 @@ CRenderableObjectsManager::~CRenderableObjectsManager(){}
 
 void CRenderableObjectsManager::Update(float ElapsedTime)
 {
-	for (int i = 0; i<m_ResourcesVector.size(); ++i)
+	for (size_t i = 0; i<m_ResourcesVector.size(); ++i)
 	{
 		m_ResourcesVector[i]->Update(ElapsedTime);
 	}
 }
 void CRenderableObjectsManager::Render(CRenderManager *RM)
 {
-	for (int i = 0; i<m_ResourcesVector.size(); ++i)
+	for (size_t i = 0; i<m_ResourcesVector.size(); ++i)
 	{
 		m_ResourcesVector[i]->Render(RM);
 	}
